Separate the template parsing from the evaluation
 - improve and harmonize error reporting

 - fix bug with empty variables, and use explicit 'tostring' (*)

Separating loading and interpretation (load() and xpcall()) enables
multiple evaluation without re-parsing the template every time.
Also, syntax errors can be caught at loading time.

The API is bit ackward though, because the evaluation environment is
bound at loading time, not evaluation time (because of Lua's load()).

(*)
This fixes the case of a user template like "$()", and also
solves the cases when the expression results in a value which
cannot be implicitly converted to string, as in "$(false)"