Refactor, improve internal API, fix table expansion
- functions to expand, load and evaluate the template

- allow the user to customize the to-string conversion,
  with the optional 'mytostring' entry in the env

- respect a newline at the end of the given template, if
  any

- fix table expansion to happen only when the token is alone
  on the line, and make the pattern greedy.

At the moment the '${<table>}' token is meant to expand only
when it is the sole non-blank string in the line, in the
template source. This commit fixes cases like

${<table>} <text>

where table expansion would occurr anyway, dropping
<text> completely (for this example the expected behavior
is instead to copy the line as it is, because it is not
really a match of any rule).

Also, the pattern match for the table-expansion argument
is now greedy, again because it is meant to be the only
token on the line. This is however controversial, as it
matches things like

${aa} ${bb}

with "aa} ${bb", which is 99% not what the user intended.
On the other hand, it allows for cases like

${ lookup["${}"] }

to work.

Adapted the tests and the samples to the changes in the API
(essentially 'parse' replaced by 'tload').