Improve on error reporting
Essentially, include all the stacktrace entries up to the call
to 'xpcall', instead of including only those referring to a
line of the user template.

This makes the error message more clear in case the error is
happening inside a function which is defined outside of the
template (although invoked from within the template). This
case is quite common.