Better constrain the allowed arguments for template/table inclusion
By replacing the any-mathing pattern ".".

I now use the explicit pattern '[%w_]' (alphanumeric characters
and underscore), to extract the name in a template inclusion,
as in "$<name>".

This more restrictive pattern rejects nasty cases like

$<aaa$<bb>ccc>

or even

$<inc1> $<inc2>

In the latter, the previous pattern (even though it is not greedy!)
will match "inc1> $<inc2".

I think this limitation is acceptable as it just requires any
included template to be given a name that is a valid identifier.

On the other hand, the same limitation for table expansion would be
excessive, as it would prevent using _expressions_ that resolve
to a table value. That is, we want cases like "${f()}" or
"${var.nested}" to be valid, obviously.

Therefore, the pattern now only rejects braces '{' '}' to be robust
against the case

${aa} ${bb}

where, once again, the pattern "(.-)$" will match "aa} ${bb".

In fact, this prevents other very specific cases which however are valid
Lua, like

${ table["aa{"] }

or simply

${ fun("{") }

But for now I do not care. TODO: add in the documentation