Add feature of template inclusion
New syntax "$<..>" to include a non-evaluated template, for
template code reuse. Added dedicated test module.

Error tracing works by storing not only the line at which a
template is included, but especially the first line of the
expanded code that comes from the included template.
Error tracing is used for both loading and evaluation errors.