Convert to string the error object in the handler
Because the error object returned by xpcall is not guaranteed to
be a string, but we can do something only when it is a string.
Thus, unconditionally convert it before matching the regex.

This change fixes those corner cases where the object is indeed
not a string, and thus the regex match causes an error in the
error handler itself, causing recursion and stack overflow.